+++ .gitignore
@@ -16,3 +16,4 @@ bin/
 /text-ui-test/ACTUAL.txt
 text-ui-test/EXPECTED-UNIX.TXT
 text-ui-test/data/duke.txt
+data/duke.txt
