--- data/duke.txt
@@ -1,2 +1,0 @@
-deadline return book /by 2022-08-28
-unmark
