+++ src/main/java/Duke/Storage.java
@@ -1,3 +1,5 @@
+package Duke;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.nio.file.Files;
