+++ src/main/java/Duke/Task/EventTask.java
@@ -1,25 +1,31 @@
+package Duke.Task;
+
+import Duke.DukeException;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-public class DeadlineTask extends Task {
-    protected LocalDate by;
-    public DeadlineTask(String description) throws DateTimeParseException, DukeException {
+public class EventTask extends Task {
+    protected LocalDate at;
+
+    public EventTask(String description) throws DateTimeParseException, DukeException {
         super();
         this.commandString = description;
-        description = description.substring(9);
-        String[] split = description.split(" /by ");
+        description = description.substring(6);
+        String[] split = description.split(" /at ");
         if (split.length < 2) {
             throw new DukeException("Deadline time (indicated by /by separator) is missing.");
         } else if (split.length > 2) {
             throw new DukeException("Multiple usage of /by separator is not allowed.");
         }
         this.description = split[0];
-        this.by = LocalDate.parse(split[1]);
+        this.at = LocalDate.parse(split[1]);
     }
 
     @Override
     public String toString() {
-        return "[D]" + super.toString() + " (by: " + by.format(DateTimeFormatter.ofPattern("MMM-dd-yyyy")) + ")";
+        return "[E]" + super.toString() + " (at: " + at.format(DateTimeFormatter.ofPattern("MMM-dd-yyyy")) + ")";
     }
+
 }
