+++ src/main/java/Duke/Task/Task.java
@@ -1,3 +1,5 @@
+package Duke.Task;
+
 abstract public class Task {
     protected String description;
     protected String commandString;
