+++ src/main/java/Duke/TaskList.java
@@ -1,3 +1,7 @@
+package Duke;
+
+import Duke.Task.Task;
+
 import java.util.ArrayList;
 
 /**
