+++ text-ui-test/EXPECTED.TXT
@@ -1,6 +1,6 @@
 Hello from
- ____        _
-|  _ \ _   _| | _____
+ ____        _        
+|  _ \ _   _| | _____ 
 | | | | | | | |/ / _ \
 | |_| | |_| |   <  __/
 |____/ \__,_|_|\_\___|
@@ -13,7 +13,7 @@ Hello from
 
     ____________________________________________________________
     Here are the tasks in your list:
-
+    
     1. [T][ ] borrow book
     ____________________________________________________________
 
@@ -41,7 +41,7 @@ Hello from
 
     ____________________________________________________________
     Here are the tasks in your list:
-
+    
     1. [T][ ] borrow book
     2. [D][ ] return book (by: Aug-28-2022)
     3. [E][X] project meeting (at: Sep-21-2022)
